+++ frontend/.storybook/main.js
@@ -3,13 +3,13 @@ const config = {
   stories: ["../src/**/*.mdx", "../src/**/*.stories.@(js|jsx|mjs|ts|tsx)"],
 
   addons: [
-    "@storybook/addon-essentials",
     "@storybook/addon-onboarding",
     "@chromatic-com/storybook",
-    "@storybook/experimental-addon-test",
+    "@storybook/addon-vitest",
     "storybook-addon-remix-react-router",
     "@storybook/addon-links",
     "@storybook/addon-a11y",
+    "@storybook/addon-docs",
   ],
 
   framework: {