+++ frontend/src/features/TransientOtp/components/OtpVerification.tsx
@@ -73,6 +73,7 @@ export default function OtpVerification({
   showTryAnotherWay = true,
   resetAttempts,
   otpExpiry = null,
+  onCancel,
 }: OtpVerificationProps) {
   const { language } = useParams();
   const [time, setTime] = useState(initialTime);
@@ -214,7 +215,7 @@ export default function OtpVerification({
       <GcdsContainer>
         <GcdsHeading tag="h1" lang={language}>
           {userMfaType === FLOW_TYPES.email
-            ? t("Verification.checkYourEmail")
+            ? t("CheckYourEmail.checkYourEmail")
             : t("Verification.checkYourPhone")}
         </GcdsHeading>
 
@@ -224,7 +225,7 @@ export default function OtpVerification({
 
             <GcdsGrid
               columns={
-                showTryAnotherWay ? "max-content max-content" : "max-content"
+                showTryAnotherWay ? "max-content" : "max-content max-content"
               }
               gap="200"
             >
@@ -249,17 +250,31 @@ export default function OtpVerification({
                 >
                   {t("Verification.chooseDifferentMethod")}
                 </GcdsButton>
-              ) : null}
+              ) : (
+                <GcdsButton
+                  buttonRole="secondary"
+                  style={{ width: "fit-content" }}
+                  onGcdsClick={(ev) => {
+                    ev.preventDefault();
+                    onCancel();
+                  }}
+                >
+                  {t("Verification.cancel")}
+                </GcdsButton>
+              )}
             </GcdsGrid>
           </>
         ) : (
           <>
+            {userMfaType === FLOW_TYPES.email && (
+              <GcdsText>{t("CheckYourEmail.toChangeYourPassword")}</GcdsText>
+            )}
             <GcdsText>
               {userMfaType === FLOW_TYPES.voice
                 ? t("Verification.voiceCodeSent")
                 : userMfaType === FLOW_TYPES.sms
                   ? t("Verification.smsCodeSent")
-                  : t("Verification.emailCodeSent")}
+                  : t("CheckYourEmail.emailCodeSent")}
               &nbsp;
               <strong>{userSelectedMfaFactor.destination}</strong>
             </GcdsText>
@@ -268,7 +283,7 @@ export default function OtpVerification({
                 ? t("Verification.callMayTakeMinutes")
                 : userMfaType === FLOW_TYPES.sms
                   ? t("Verification.smsMayTakeMinutes")
-                  : t("Verification.emailMayTakeMinutes")}
+                  : t("CheckYourEmail.emailMayTakeMinutes")}
             </GcdsText>
             <GcdsText>
               {t("Verification.codeExpiresIn")}{" "}
@@ -301,7 +316,7 @@ export default function OtpVerification({
 
             <GcdsGrid
               columns={
-                showTryAnotherWay ? "max-content max-content" : "max-content"
+                showTryAnotherWay ? "max-content" : "max-content max-content"
               }
               gap="200"
             >
@@ -325,7 +340,18 @@ export default function OtpVerification({
                 >
                   {t("Verification.chooseDifferentMethod")}
                 </GcdsButton>
-              ) : null}
+              ) : (
+                <GcdsButton
+                  buttonRole="secondary"
+                  style={{ width: "fit-content" }}
+                  onGcdsClick={(ev) => {
+                    ev.preventDefault();
+                    onCancel();
+                  }}
+                >
+                  {t("Verification.cancel")}
+                </GcdsButton>
+              )}
             </GcdsGrid>
           </>
         )}
@@ -351,7 +377,7 @@ export default function OtpVerification({
           >
             {userMfaType !== FLOW_TYPES.email
               ? t("Verification.requestNewCode")
-              : t("Verification.sendCodeAgain")}
+              : t("CheckYourEmail.sendCodeAgain")}
           </GcdsLink>
         )}
       </GcdsText>