+++ frontend/src/stories/Tests/Features/MFAPhoneNumber/AddMFAPhoneNumber/component/AddMFAPage.test.stories.jsx
@@ -1,4 +1,4 @@
-import { expect, userEvent, waitFor, within } from "@storybook/test";
+import { expect, userEvent, waitFor, within } from "storybook/test";
 import {
   AVAILABLE_LANGUAGES,
   FLOW_TYPES,